Move function to new_variables
